tivix/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.9.4
